% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{b.extra}
\alias{b.extra}
\title{b.extra}
\usage{
b.extra(obj)
}
\arguments{
\item{obj, }{mbSet class}
}
\value{
the otu abundance data table
}
\description{
obtain the b slot from mbSet
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
b.extra(object)
}
