% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{b}
\alias{b}
\title{b

Extract the OTU abundance matrix from the mbSet class object.}
\usage{
b(object)
}
\arguments{
\item{object}{mbSet class}
}
\value{
data.table OTU aundance matrix
}
\description{
b

Extract the OTU abundance matrix from the mbSet class object.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
b(object)
}
