% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{bNum}
\alias{bNum}
\title{bNum}
\usage{
bNum(object)
}
\arguments{
\item{object}{mbSet class}
}
\value{
OTU number
}
\description{
Calculate the OTU number from the mbSet class object.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
bNum(object)
}
