% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{bSet}
\alias{bSet}
\title{bSet}
\usage{
bSet(b, ...)
}
\arguments{
\item{b}{data.table, metabolites aundance matrix. if `rn` column is not contained
in this data.table, the `Features` parameter should be given by character vector.}

\item{...}{`Samples` or `Features`. if the `Samples` not given, the colnames of
`b` data.table will be taken as the Samples names.}
}
\value{
a `bSet` class
}
\description{
Function to return `bSet` class.
}
\examples{
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(genera)[1] <- 'rn'
bSet(b = genera)
}
