% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_mbSet}
\alias{check_mbSet}
\title{check_mbSet

`check_mbSet` check the mbSet class.}
\usage{
check_mbSet(object)
}
\arguments{
\item{object}{mbSet}
}
\value{
logical
}
\description{
check_mbSet

`check_mbSet` check the mbSet class.
}
