% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{clean_analytes}
\alias{clean_analytes}
\title{clean_analytes}
\usage{
clean_analytes(object, m_thres = 2, b_thres = 2)
}
\arguments{
\item{object, }{mbSet class}

\item{m_thres, }{integer, Removal of metabolites only measured in m_thres of samples}

\item{b_thres, }{integer, Removal of OTU only measured in b_thres of samples}
}
\description{
clean_analytes
}
