% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{cluster_jsd}
\alias{cluster_jsd}
\title{cluster_jsd}
\usage{
cluster_jsd(dist, b, k)
}
\arguments{
\item{dist}{distance matrix, if miss distance matrix the bacterial abundance matrix `b` was used to calculate the Jensen-Shannon divergence}

\item{b}{bacterial abundance matrix, if the distance matrix `dist` was given, it is useless}

\item{k}{number of cluster}
}
\value{
cluster vector
}
\description{
cluster the sample using the pam based the JSD
}
