% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-wgcna.R
\name{coExpress,mbSet-method}
\alias{coExpress,mbSet-method}
\title{coExpress}
\usage{
\S4method{coExpress}{mbSet}(
  object,
  power = NULL,
  powerVec = 1:30,
  threshold = 0.8,
  message = TRUE,
  ...
)
}
\value{
network
}
\description{
coExpress
}
