% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{cor.test}
\alias{cor.test}
\alias{cor.test,mbSet-method}
\title{cor.test}
\usage{
cor.test(object, parallel = FALSE, ncore = 4, ...)

\S4method{cor.test}{mbSet}(object, parallel = FALSE, ncore = 4, ...)
}
\arguments{
\item{object, }{mbSet class}

\item{parallel, }{logical}

\item{ncore, }{integer number of core}
}
\value{
the correlation data table
}
\description{
genetic methods to perform the correlation test
}
\examples{
res <- cor.test(object, method = 'spearman')
}
