% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-corr.R
\docType{methods}
\name{corr}
\alias{corr}
\title{corr}
\usage{
corr(m, b, method = "spearman", parallel = FALSE, ncore = 4)
}
\arguments{
\item{m}{mSet class}

\item{b}{bSet class}

\item{method}{character default is spearman}

\item{parallel}{logical}

\item{ncore}{integer number of core,default is 4}
}
\value{
the correlation data table
}
\description{
Calculating spearman correlation between each OTU and
each metabolites.
}
\examples{
library(data.table)
path <- system.file('extdata','metabolites_and_genera.rda',package = 'mbOmic')
load(path)
names(genera)[1] <- 'rn'
names(metabolites)[1] <- 'rn'
b <- bSet(b = genera)
m <- mSet(m = metabolites)
res <- corr(m, b, method = 'spearman')
}
