% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{dist_jsd}
\alias{dist_jsd}
\title{dist_jsd

Calculating Jensen-Shannon divergence basing the formulate
\eqn{JSD(P || Q) =
      \sqrt{0.5 \times KLD (P||\frac{P + Q}{2}) + 0.5 \times KLD(Q||\frac{P+Q}{2})} }.}
\usage{
dist_jsd(b)
}
\arguments{
\item{b}{a `bSet` class}
}
\value{
Jensen-Shannon divergence data frame
}
\description{
dist_jsd

Calculating Jensen-Shannon divergence basing the formulate
\eqn{JSD(P || Q) =
      \sqrt{0.5 \times KLD (P||\frac{P + Q}{2}) + 0.5 \times KLD(Q||\frac{P+Q}{2})} }.


}
\examples{
data(iris)
b <- data.table::as.data.table(iris[1:6, 1:4])
b <- bSet(b = b,
          Features = letters[1:6],
          Samples = LETTERS[1:4])
dist <- dist_jsd(b)
}
