% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{enterotyping}
\alias{enterotyping}
\title{enterotyping}
\usage{
enterotyping(b, cluster, threshold = 0.02)
}
\arguments{
\item{b}{bacterial abundance matrix}

\item{cluster}{cluster vector}

\item{threshold}{abundance threshold}
}
\value{
list
}
\description{
identifing enterotype basing the OTU abundance using
the cluster anlysis refer to the XXX works.
}
\examples{
dat <- read.delim('http://enterotypes.org/ref_samples_abundance_MetaHIT.txt')
}
