% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{estimate_k}
\alias{estimate_k}
\title{estimate_k}
\usage{
estimate_k(b, verK = 2:10)
}
\arguments{
\item{b}{bacterial abundance matrix}

\item{verK}{number vector of cluster}
}
\value{
list
}
\description{
To estimate the optimal cluster number ,
`estimate_k` takes advantage of two measures,
Calinski-Harabasz (CH) Index and silhouette coefficients.
}
\examples{
data(iris)
b <- bSet(b = data.table::as.data.table(t(iris[1:6, 1:4])),
          Features = letters[1:4],
          Samples = LETTERS[1:6])
ret <- estimate_k(b = b, 2:3)
}
