% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{mNum}
\alias{mNum}
\alias{bNum}
\alias{m}
\alias{b}
\alias{mbSamples}
\title{mNum}
\usage{
mNum(object)

bNum(object)

m(object)

b(object)

mbSamples(object)
}
\description{
mNum

bNum

m

b

mbSamples
}
