% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{features}
\alias{features}
\alias{features<-}
\title{features}
\usage{
features(object)
features(object) <- value

features(object)
features(object) <- value
}
\arguments{
\item{object}{a `Set` object}

\item{value}{character vector}
}
\value{
`features` return a character vector or update the features names

`features` return a character vector or update the features names
}
\description{
`features` get or set features names vector.

`features` get or set features names vector.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(metabolites)[1] <- 'rn'
m <- mSet(m = metabolites)
features(m)
}
