% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{m.extra}
\alias{m.extra}
\title{m.extra}
\usage{
m.extra(obj)
}
\arguments{
\item{obj, }{mbSet class}
}
\value{
the metabolites abundance data table
}
\description{
obtain the m slot from mbSet
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
m.extra(object)
}
