% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{m}
\alias{m}
\title{m

Extract the metabolites abundance matrix from the mbSet class object.}
\usage{
m(object)
}
\arguments{
\item{object}{mbSet class}
}
\value{
data.table metabolites aundance matrix
}
\description{
m

Extract the metabolites abundance matrix from the mbSet class object.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
m(object)
}
