% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{mb_initialize}
\alias{mb_initialize}
\title{mb_initialize

initialized function for `mSet` and `bSet` classs.}
\usage{
mb_initialize(type, dt, Samples, Features)
}
\arguments{
\item{type}{character `mSet` or `bSet`}

\item{dt}{data.table abundance matrix}

\item{Samples}{character samples names}

\item{Features}{character Features names}
}
\value{
object
}
\description{
mb_initialize

initialized function for `mSet` and `bSet` classs.
}
