% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{mbSamples}
\alias{mbSamples}
\title{mbSamples
extract the samples names from a mbSet class object.}
\usage{
mbSamples(object)
}
\arguments{
\item{object}{mbSet class}
}
\value{
samples number
}
\description{
mbSamples
extract the samples names from a mbSet class object.
}
\examples{

library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
mbSamples(object)
}
