% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mbSet-class}
\alias{mbSet-class}
\title{mbSet}
\description{
`mbSet` is the main S4 class of mbOmic. It contains the
 metabolites (`m`) and OTU (`b`) abundance matrix.
 The `anno.m` and `anno.b` are the annotation of the `m`
 and `b`, respectially, which need to implement on the later.
}
\section{Slots}{

\describe{
\item{\code{b}}{OTU, column response to samples, row response to microbiota}

\item{\code{m}}{metabolites, column response to samples, row response to metabolites}

\item{\code{anno.m}}{metabolites features}

\item{\code{anno.b}}{microbiota features}

\item{\code{samples}}{samples names}
}}

