% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mbSet}
\alias{mbSet}
\title{mbSet}
\usage{
mbSet(m, b, anno.m, anno.b, m.f, b.f)
}
\arguments{
\item{m}{the dataframe of metabolites abundance profile}

\item{b}{the dataframe of OTU abundance profile}

\item{anno.m}{dataframe for annotaing metabolites}

\item{anno.b}{dataframe for annotating OTU}

\item{m.f}{character th column contains metabolite ID}

\item{b.f}{character the column contains OTU ID}
}
\value{
mbSet object
}
\description{
create a mbSet class.
}
\examples{
path <- system.file('extdata',
                    'metabolites_and_genera.rda',
                     package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
}
