% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{mNum}
\alias{mNum}
\title{mNum}
\usage{
mNum(object)
}
\arguments{
\item{object}{mbSet class}
}
\value{
metabolites number
}
\description{
Calculate the metabolites number from the mbSet class object.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
mNum(object)
}
