% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{mSet-class}
\alias{mSet-class}
\title{mSet}
\value{
S4 class
}
\description{
`mSet` is a S4 class extended from the virtual `Set` used as
object to store metabolites abundance matrix.
}
\section{Slots}{

\describe{
\item{\code{Samples}}{character a character vector contains the samples}

\item{\code{Features}}{character a character vector contains the features}

\item{\code{dt}}{data.table metabolties abundance matrix}
}}

\examples{
m.path <- system.file("extdata", "metabolites_and_genera.rda", package = "mbOmic")
load(m.path)
names(metabolites)[1] <- 'rn'
bSet(b = metabolites)
}
