% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{mSet}
\alias{mSet}
\title{mSet}
\usage{
mSet(m, ...)
}
\arguments{
\item{m}{data.table, metabolites aundance matrix. if `rn` column is not contained
in this data.table, the `Features` parameter should be given by character vector.}

\item{...}{`Samples` or `Features`. if the `Samples` not given, the colnames of
`m` data.table will be taken as the Samples names.}
}
\value{
a `mSet` class
}
\description{
Function to return `mSet` class.
}
\examples{
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(metabolites)[1] <- 'rn'
mSet(m = metabolites)
}
