% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nb.extra}
\alias{nb.extra}
\title{nb}
\usage{
nb.extra(obj)
}
\arguments{
\item{obj, }{mbSet class}
}
\value{
number of otu
}
\description{
nb
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
nb.extra(object)
}
