% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{ncol,Set-method}
\alias{ncol,Set-method}
\title{ncol}
\usage{
\S4method{ncol}{Set}(x)
}
\arguments{
\item{x}{a `Set` Object}
}
\value{
an `integer` of length 1
}
\description{
ncol method for `Set` to obtain the number of samples.
nrow and ncol return the number of features or samples of a `Set` class.
}
