% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{nm.extra}
\alias{nm.extra}
\title{nm}
\usage{
nm.extra(obj)
}
\arguments{
\item{obj, }{mbSet class}
}
\value{
number of metabolites
}
\description{
nm
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
nm.extra(object)
}
