% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{nrow,Set-method}
\alias{nrow,Set-method}
\title{nrow}
\usage{
\S4method{nrow}{Set}(x)
}
\arguments{
\item{x}{a `Set` Object}
}
\value{
an `integer` of length 1
}
\description{
nrow method for `Set` to obtain the number of features.
nrow and ncol return the number of features or samples of a `Set` class.
}
