% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(
    net = net,
    corr = corr，
    centrality_degree_mode = 'out',
    threshold = 0.9
)
}
\arguments{
\item{net, }{the result of function `wgcna`}

\item{corr, }{the result of function `corr.test`}

\item{centrality_degree_mode, }{the mode for centrality_degress function}

\item{threshold, }{the nodes whose centrality degree greater than the theshold will show labels}
}
\value{
ggraph
}
\description{
plot network
}
\examples{
\dontrun{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
res <- corr(object, method = 'spearman')
net <- wgcna(object, minN = 2, power = 9,message = FALSE)
plot_network(net, res[abs(rho)>=0.85])
}
}
\author{
Congcong Gong
}
