% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{print.verCHI}
\alias{print.verCHI}
\title{print.verCHI}
\usage{
\method{print}{verCHI}(x, ..., verbose = TRUE, plotting = TRUE, cluster)
}
\arguments{
\item{x}{estimate_k output}

\item{...}{print}

\item{verbose}{logical}

\item{plotting}{logical}

\item{cluster}{cluster result}
}
\value{
no return
}
\description{
print the optialmal numer of cluster, max CHI, and Silhouette
and plot the Calinski-Harabasz (CH) Index and silhouette coefficients simultaneously.
}
\examples{
data(iris)
ret <- estimate_k(b = t(iris[,1:4]), 1:10)
ret
}
