% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{samples.extra}
\alias{samples.extra}
\title{samples}
\usage{
samples.extra(obj)
}
\arguments{
\item{obj, }{mbSet class}
}
\value{
a vector containing sample names
}
\description{
obtain the samples
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
samples.extra(object)
}
