% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{samples}
\alias{samples}
\alias{samples<-}
\title{samples}
\usage{
samples(object)
samples(object) <- value

samples(object)
samples(object) <- value
}
\arguments{
\item{object}{a `Set` object}

\item{value}{character vector}
}
\value{
`samples` return a character vector or update the samples names

`samples` return a character vector or update the samples names
}
\description{
`samples` get or set samples names vector.

`samples` get or set samples names vector.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(metabolites)[1] <- 'rn'
m <- mSet(m = metabolites)
features(m)
}
