% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Set-class}
\alias{Set-class}
\title{Set}
\value{
no return
}
\description{
`Set` is a virtual class as the base class to
extend the `mSet` and `bSet`.
}
\section{Slots}{

\describe{
\item{\code{Samples}}{character a character vector contains the samples}

\item{\code{Features}}{character a character vector contains the features}

\item{\code{dt}}{data.table}
}}

\examples{
##`Set` is virtual class.
}
