% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{show,Set-method}
\alias{show,Set-method}
\title{show}
\usage{
\S4method{show}{Set}(object)
}
\arguments{
\item{object}{a `Set` class}
}
\value{
none return but print the main info of `Set` class
}
\description{
show method for the `Set` class.
}
