% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{[.Set}
\alias{[.Set}
\title{subset_Set}
\usage{
\method{[}{Set}(x, i, j)
}
\arguments{
\item{x}{a `mSet` or `bSet` class}

\item{i}{the data.table i}

\item{j}{the data.table j}
}
\value{
return a subset object
}
\description{
subset `Set` class.
}
