% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmh-class.R
\docType{class}
\name{vmh-class}
\alias{vmh-class}
\title{vmh}
\value{
vmh class
}
\description{
`vmh` is a API for vmh
}
\section{Slots}{

\describe{
\item{\code{type}}{a character, "microbe", "metabolites", "gene", or "food items"}

\item{\code{net}}{a data.table class contain the network of microbe, metabolites, gene, and food items}
}}

\examples{
b.path <- system.file("extdata", "metabolites_and_genera.rda", package = "mbOmic")
load(b.path)
names(genera)[1] <- 'rn'
bSet(b = genera)
}
