% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{wgcna,mbSet-method}
\alias{wgcna,mbSet-method}
\title{wgcna}
\usage{
\S4method{wgcna}{mbSet}(
  object,
  group,
  minN,
  power = NULL,
  powers = 1:30,
  threshold.d = 0.05,
  threshold = 0.8,
  message = FALSE,
  ...
)
}
\value{
network
}
\description{
wgcna
}
