% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{wgcna}
\alias{wgcna}
\title{wgcna}
\usage{
wgcna(
  object,
  group,
  minN = 2,
  power = NULL,
  powers = 1:30,
  threshold.d = 0.05,
  threshold = 0.8,
  message = FALSE,
  ...
)
}
\arguments{
\item{object, }{mbSet class}

\item{minN, }{integer the minimum number of sample}

\item{power, }{integer, if the pickSoftThreshold function (WGCNA) can find appropriate power, this param is invalid}
}
\value{
network
}
\description{
wgcna
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
net <- wgcna(object, minN = 2, power = 9,message = FALSE)
}
\author{
Congcong Gong
}
