% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilites.R
\name{clean_analytes}
\alias{clean_analytes}
\title{clean_analytes}
\usage{
clean_analytes(object, m_thres = 2, b_thres = 2)
}
\arguments{
\item{object, }{mbSet class}

\item{m_thres, }{integer, Removal of metabolites only measured in m_thres of samples}

\item{b_thres, }{integer, Removal of OTU only measured in b_thres of samples}
}
\value{
mbSet object
}
\description{
Remove unqualified metabolites and OTU which only
measured in few samples.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
clean_analytes(object)
}
