% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-preprocess.R
\name{clean_analytes}
\alias{clean_analytes}
\title{clean_analytes}
\usage{
clean_analytes(object, fea_num = 2)
}
\arguments{
\item{object}{a `bSet` or `mSet` class}

\item{fea_num, }{integer, Removal of features only measured in `fea_num` of samples}
}
\value{
return a quantified `bSet` or `mSet` object
}
\description{
Remove unqualified metabolites and OTU which only
measured in few samples.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(genera)[1] <- 'rn'
names(metabolites)[1] <- 'rn'
b <- bSet(b = genera)
m <- mSet(m = metabolites)
clean_analytes(b)
}
