% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{clean_analytes}
\alias{clean_analytes}
\alias{clean_analytes,mbSet-method}
\title{clean_analytes}
\usage{
clean_analytes(object, m_thres = 2, b_thres = 2)

\S4method{clean_analytes}{mbSet}(object, m_thres = 2, b_thres = 2)
}
\arguments{
\item{object, }{mbSet class}

\item{m_thres, }{integer, Removal of metabolites only measured in m_thres of samples}

\item{b_thres, }{integer, Removal of OTU only measured in b_thres of samples}
}
\value{
mbSet
}
\description{
Removal of analytes only measured in few samples
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
mb <- clean_analytes(object, 2, 2)
}
