% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{cluster_jsd}
\alias{cluster_jsd}
\title{cluster_jsd}
\usage{
cluster_jsd(dist, b, k)
}
\arguments{
\item{dist}{distance matrix, if miss distance matrix the
bacterial abundance matrix `b` was used to calculate the
Jensen-Shannon divergence}

\item{b}{bacterial abundance matrix, if the distance matrix
`dist` was given, it is useless}

\item{k}{number of cluster}
}
\value{
cluster vector
}
\description{
`cluster_jsd` cluster the sample using the pam based the JSD.
}
\examples{

data(iris)
## using the matrix to cluster samples.
b <- bSet(b = data.table::as.data.table(t(iris[1:6, 1:4])),
          Features = letters[1:4],
          Samples = LETTERS[1:6])
res_cluster <- cluster_jsd(b=b, k = 3)
## or using the resoult of dist_jsd
## dist <- dist_jsd(iris[, 1:4])
## res_cluster <- cluster_jsd(dist = iris[,1:4], k = 3)
table(res_cluster, iris[1:6, 5])

}
