% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-corr.R
\docType{methods}
\name{corr,mbSet-method}
\alias{corr,mbSet-method}
\title{corr}
\usage{
\S4method{corr}{mbSet}(object, parallel = FALSE, ncore = 4, ...)
}
\description{
genetic methods to perform the correlation test
}
