% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-corr.R
\docType{methods}
\name{corr}
\alias{corr}
\title{corr}
\usage{
corr(object, parallel = FALSE, ncore = 4, ...)
}
\arguments{
\item{object, }{mbSet class}

\item{parallel, }{logical}

\item{ncore, }{integer number of core,default is 4}
}
\value{
the correlation data table
}
\description{
corr
}
\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
res <- corr(object, method = 'spearman')
}
