% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{dist_jsd}
\alias{dist_jsd}
\title{dist_jsd}
\usage{
dist_jsd(b)
}
\arguments{
\item{b}{bacterial abundance matrix}
}
\value{
Jensen-Shannon divergence data frame
}
\description{
calculate Jensen-Shannon divergence
}
