% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{enterotyping}
\alias{enterotyping}
\title{enterotyping}
\usage{
enterotyping(b, cluster, threshold = 0.25)
}
\arguments{
\item{b}{bacterial abundance matrix}

\item{cluster}{cluster vector}

\item{threshold}{abundance threshold}
}
\value{
list
}
\description{
identify enterotype
}
