% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{estimate_k}
\alias{estimate_k}
\title{estimate_k}
\usage{
estimate_k(b, verK = 2:10)
}
\arguments{
\item{b}{bacterial abundance matrix}

\item{verK}{number vector of cluster}
}
\value{
list
}
\description{
To estimate the optimal cluster number ,
`estimate_k` takes advantage of two measures,
Calinski-Harabasz (CH) Index and silhouette coefficients.
}
\examples{
data(iris)
ret <- estimate_k(b = t(iris[, 1:4]), 1:10)
}
