% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{mbSet}
\alias{mbSet}
\title{mbSet}
\usage{
mbSet(m, b, anno.m, anno.b, m.f, b.f)
}
\arguments{
\item{m, }{the dataframe of metabolites abundance profile}

\item{b, }{the dataframe of OTU abundance profile}
}
\value{
mbSet object
}
\description{
mbSet

create a mbSet class
}
\section{Slots}{

\describe{
\item{\code{b}}{abundance of OTU, column response to samples, row response to microbiota}

\item{\code{m}}{abundance abundance, column response to samples, row response to metabolites}

\item{\code{anno.m}}{metabolites features}

\item{\code{anno.b}}{microbiota features}

\item{\code{samples}}{samples names}
}}

\examples{
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
}
