% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/functions.R
\name{mbSet}
\alias{mbSet}
\title{An S4 class to represent a bank account.}
\usage{
mbSet(m, b, pData = NULL)

mbSet(m, b, pData = NULL)
}
\arguments{
\item{m, }{the dataframe of metabolites abundance profile}

\item{b, }{the dataframe of OTU abundance profile}
}
\value{
mbSet class
}
\description{
obtain the m express matrix from mbSet
create the mbSet class
}
\section{Slots}{

\describe{
\item{\code{b}}{abundance of OTU}

\item{\code{m}}{abundance abundance}

\item{\code{pData}}{phenodata}
}}

\examples{
library(data.table)
path <- system.file('data',package = 'mbOmic')
load(file.path(path,'metabolites_and_genera.rda'))
object <-
       mbSet(
             m = metabolites,
             b = genera
             )
}
