% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m-wgcna.R
\name{pickST}
\alias{pickST}
\title{pickST}
\usage{
pickST(m, threshold.d = 0.05, threshold = 0.8, plot = TRUE, powers = NULL)
}
\value{
power
}
\description{
Automatic network construction and module detection by one-step method
}
