% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coExpress}
\alias{plot_coExpress}
\title{plot_coExpress}
\usage{
plot_coExpress(net)
}
\arguments{
\item{net}{output of coExpress function}
}
\value{
ploting
}
\description{
Plotting the dendro of metabolites modules.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(metabolites)[1] <- 'rn'
m <- mSet(m = metabolites)
net <- coExpress(m, minN = 2, power = 9, message = FALSE)
plot_coExpress(net)

}
