% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(
    net,
    corr,
    centrality_degree_mode = 'out',
    threshold = 0.9,
    show_text = TRUE
)
}
\arguments{
\item{net}{result of function `wgcna`}

\item{corr}{result of function `corr.test`}

\item{centrality_degree_mode}{the mode for centrality_degress function}

\item{threshold, }{nodes whose centrality degree greater than the theshold will show labels}

\item{show_text}{logical whether add the nodes labels}
}
\value{
ggraph
}
\description{
plotting the network of metabolites and OTU. Orange nodes represent the OTU, while the other color represent the metabolite. Same color metabolites nodes are constructed in the same modules.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
mb <-
     mbSet(
           m = metabolites, b = genera
           )
           mb
## Samples( 12 ):  BS1 BS2 BS3 BS4 BS5 BS6 SS1 SS2 SS3 SS4 SS5 SS6
## number of OTU:  18
## number of metabolites:  247
net <- coExpress(mb,message = TRUE,threshold.d = 0.02, threshold = 0.8, power = 9)
spearm <- corr(mb, method = 'spearman')
plot_network(net, spearm[abs(rho) >= 0.75 & p <= 0.05], threshold = 0.75, show_text = FALSE)
}
\author{
Congcong Gong
}
