% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network}
\alias{plot_network}
\title{plot_network}
\usage{
plot_network(
    net,
    corr,
    centrality_degree_mode = 'out',
    threshold = 0.9,
    show_text = TRUE
)
}
\arguments{
\item{net}{result of function `wgcna`}

\item{corr}{result of function `corr.test`}

\item{centrality_degree_mode}{the mode for centrality_degress function}

\item{threshold, }{nodes whose centrality degree greater than the theshold will show labels}

\item{show_text}{logical whether add the nodes labels}
}
\value{
ggraph
}
\description{
plotting the network of metabolites and OTU. Orange nodes represent the OTU, while the other color represent the metabolite. Same color metabolites nodes are constructed in the same modules.
}
\examples{
library(data.table)
path <- system.file('extdata', 'metabolites_and_genera.rda', package = 'mbOmic')
load(path)
names(genera)[1] <- 'rn'
names(metabolites)[1] <- 'rn'
b <- bSet(b = genera)
m <- mSet(m = metabolites)
res <- corr(m, b, method = 'spearman')
net <- coExpress(m, minN = 2, power = 9, message = FALSE)
plot_network(net, res[abs(rho)>=0.85])
}
\author{
Congcong Gong
}
