% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterotype.R
\name{print.verCHI}
\alias{print.verCHI}
\title{print.verCHI}
\usage{
\method{print}{verCHI}(verCHI, verbose = TRUE, plotting = TRUE, cluster)
}
\arguments{
\item{verCHI}{estimate_k output}

\item{verbose}{logical}
}
\value{
no return
}
\description{
print verCHI
}
\examples{
data(iris)
ret <- estimate_k(b = t(iris[,1:4]), 1:10)
ret
}
